BD changes

+++ src/main/java/com/solvd/ATMproject/models/User.java
@@ -7,7 +7,7 @@ public class User {
     int idUser;
     String firstName;
     String lastName;
-    int cardNumber;
+    long cardNumber;
     int PIN;
     int phoneNumber;
 }
